requirements.txt
seleniumbase
beautifulsoup4
requests
python-dotenv